@@ -1,14 +1,17 @@
 module github.com/chuckha/aoc
 
-go 1.17
+go 1.21
 
-require github.com/pkg/errors v0.9.1
+require (
+	github.com/mitchellh/go-wordwrap v1.0.1
+	gitlab.com/tozd/go/errors v0.8.0
+	jaytaylor.com/html2text v0.0.0-20211105163654-bc68cce691ba
+)
 
 require (
 	github.com/mattn/go-runewidth v0.0.9 // indirect
-	github.com/mitchellh/go-wordwrap v1.0.1 // indirect
 	github.com/olekukonko/tablewriter v0.0.5 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 	github.com/ssor/bom v0.0.0-20170718123548-6386211fdfcf // indirect
 	golang.org/x/net v0.0.0-20211118161319-6a13c67c3ce4 // indirect
-	jaytaylor.com/html2text v0.0.0-20211105163654-bc68cce691ba // indirect
 )
