@@ -4,7 +4,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/pkg/errors"
+	"gitlab.com/tozd/go/errors"
 )
 
 type Answer struct {
