@@ -6,25 +6,24 @@ import (
 	"github.com/chuckha/aoc/internal/core/domain"
 )
 
-type questionCache interface {
-	Insert(*domain.Question) error
-	Get(year, day int) (*domain.Question, error)
+type cachedReader interface {
+	Get(ctx context.Context, year, day int) (*domain.Question, error)
+	Delete(year, day int) error
 }
 
-type questionInterface interface {
-	Get(ctx context.Context, year, day int) (*domain.Question, error)
+type writer interface {
 	Submit(ctx context.Context, year, day int, answer *domain.Answer) ([]byte, error)
 }
 
 type Service struct {
-	QuestionCache     questionCache
-	QuestionInterface questionInterface
+	reader cachedReader
+	writer writer
 }
 
-func NewService(qc questionCache, qi questionInterface) *Service {
+func NewService(r cachedReader, w writer) *Service {
 	return &Service{
-		QuestionCache:     qc,
-		QuestionInterface: qi,
+		reader: r,
+		writer: w,
 	}
 }
 
@@ -43,30 +42,9 @@ func (s *Service) GetDescription(ctx context.Context, in *GetDescriptionInput) (
 
 func (s *Service) getQuestion(ctx context.Context, year, day int, force bool) (*domain.Question, error) {
 	if force {
-		q, err := s.QuestionInterface.Get(ctx, year, day)
-		if err != nil {
-			return nil, err
-		}
-		if err := s.QuestionCache.Insert(q); err != nil {
-			return nil, err
-		}
-		return q, nil
+		s.reader.Delete(year, day)
 	}
-
-	q, err := s.QuestionCache.Get(year, day)
-	if err != nil {
-		return nil, err
-	}
-	if q == nil {
-		q, err = s.QuestionInterface.Get(ctx, year, day)
-		if err != nil {
-			return nil, err
-		}
-		if err := s.QuestionCache.Insert(q); err != nil {
-			return nil, err
-		}
-	}
-	return q, nil
+	return s.reader.Get(ctx, year, day)
 }
 
 func (s *Service) GetInput(ctx context.Context, year, day int) ([]byte, error) {
@@ -78,9 +56,9 @@ func (s *Service) GetInput(ctx context.Context, year, day int) ([]byte, error) {
 }
 
 func (s *Service) SubmitAnswer(ctx context.Context, year, day, level int, answer string) ([]byte, error) {
-	aswr, err := domain.NewAnswer(level, answer)
+	ans, err := domain.NewAnswer(level, answer)
 	if err != nil {
 		return nil, err
 	}
-	return s.QuestionInterface.Submit(ctx, year, day, aswr)
+	return s.writer.Submit(ctx, year, day, ans)
 }
